+++ 1-foundation/practical-logger-types/src/index.ts
@@ -86,9 +86,13 @@ export interface BaseInternalLoggerState {
 }
 
 // suggested creation params
-export interface LoggerCreationParams {
+export interface BaseSinkOptions {
+	sink?: LogSink
+}
+export interface LoggerCreationParams<SinkOptions extends BaseSinkOptions = {}> {
 	name?: string
 	suggestedLevel?: LogLevel // the code is free to suggest a default level, but can expect it to be dynamically overriden (see Universal Debug API)
 	forcedLevel?: LogLevel // use only if you provide your own mechanism for dynamically setting the level
 	commonDetails?: Readonly<LogDetails>
+	sinkOptions?: SinkOptions // options specifically targeted at the sink, platform dependent
 }