+++ 1-foundation/practical-logger-types/src/index_spec.ts
@@ -1,0 +1,21 @@
+import {
+	LogLevel,
+	LogDetails,
+	LogPrimitive,
+	Logger,
+	LogPayload,
+	LogSink,
+	BaseInternalLoggerState,
+	LoggerCreationParams,
+} from '.'
+
+describe('@offirmo/practical-logger-types', function () {
+
+	describe('types', function () {
+
+		it('should compile with no errors', () => {
+			// that's it
+			const log_level: LogLevel = 'warn'
+		})
+	})
+})